requirements.txt
psycopg[binary]
pandas
numpy
matplotlib
datetime
optuna
python_dotenv